+++ tests/common/testdata.hpp
@@ -4,6 +4,8 @@
 #ifndef TESTDATA_HPP
 #define TESTDATA_HPP
 
+#include "util/precision.hpp"
+
 #include <functional>
 
 namespace qsr {
@@ -11,9 +13,9 @@ namespace qsr {
 // Number of data points
 constexpr int test_data_length = 128;
 
-void generate_test_data(float **&X, float *&y, std::function<float(float&)> ground_truth);
+void generate_test_data(double **&X, double *&y, std::function<double(double&)> ground_truth);
 
-void delete_test_data(float **&X, float *&y);
+void delete_test_data(double **&X, double *&y);
 
 }
